Restricción 1 subida por dia eliminada y fallos subsecuentes solventados

+++ backend/src/main/java/com/example/application/constructionsite/ConstructionSiteService.java
@@ -53,7 +53,10 @@ public class ConstructionSiteService {
         site.setOwner(null);
         var details = site.getDetails();
         user.removeConstructionSite(site);
-        restrictionRepository.deleteAll(site.getDetails().getAllRestrictions());
+        if (!site.getDetails().getAllRestrictions().isEmpty())
+        {
+            restrictionRepository.deleteAll(site.getDetails().getAllRestrictions());
+        }
         constructionSiteDetailsRepository.delete(details);
         constructionSiteRepository.delete(site);
     }
